+++ pylint_secure_coding_standard.py
@@ -15,7 +15,9 @@
 
 """Main file for the pylint_secure_coding_standard plugin."""
 
+import operator
 import platform
+import stat
 
 import astroid
 from pylint.checkers import BaseChecker
@@ -281,6 +283,93 @@ def _is_yaml_unsafe_call(node):
 
 # ==============================================================================
 
+_unop = {'-': operator.neg, 'not': operator.not_, '~': operator.inv}
+_binop = {
+    '+': operator.add,
+    '-': operator.sub,
+    '*': operator.mul,
+    '/': operator.truediv,
+    '//': operator.floordiv,
+    '%': operator.mod,
+    '^': operator.xor,
+    '|': operator.or_,
+    '&': operator.and_,
+}
+_chmod_known_mode_values = (
+    'S_ISUID',
+    'S_ISGID',
+    'S_ENFMT',
+    'S_ISVTX',
+    'S_IREAD',
+    'S_IWRITE',
+    'S_IEXEC',
+    'S_IRWXU',
+    'S_IRUSR',
+    'S_IWUSR',
+    'S_IXUSR',
+    'S_IRWXG',
+    'S_IRGRP',
+    'S_IWGRP',
+    'S_IXGRP',
+    'S_IRWXO',
+    'S_IROTH',
+    'S_IWOTH',
+    'S_IXOTH',
+)
+
+
+def _chmod_get_mode(node):
+    """
+    Extract the mode constant of a node.
+
+    Args:
+        node (astroid.node_classes.NodeNG): an AST node
+
+    Raises:
+        ValueError: if a node is encountered that cannot be processed
+    """
+    if isinstance(node, astroid.Name) and node.name in _chmod_known_mode_values:
+        return getattr(stat, node.name)
+    if (
+        isinstance(node, astroid.Attribute)
+        and isinstance(node.expr, astroid.Name)
+        and node.attrname in _chmod_known_mode_values
+        and node.expr.name == 'stat'
+    ):
+        return getattr(stat, node.attrname)
+    if isinstance(node, astroid.UnaryOp):
+        return _unop[node.op](_chmod_get_mode(node.operand))
+    if isinstance(node, astroid.BinOp):
+        return _binop[node.op](_chmod_get_mode(node.left), _chmod_get_mode(node.right))
+
+    raise ValueError(f'Do not know how to process node: {node.repr_tree()}')
+
+
+def _chmod_has_wx_for_go(node):
+    if platform.system() == 'Windows':
+        # On Windows, only stat.S_IREAD and stat.S_IWRITE can be used, all other bits are ignored
+        return False
+
+    try:
+        modes = None
+        if len(node.args) > 1:
+            modes = _chmod_get_mode(node.args[1])
+        elif node.keywords:
+            for keyword in node.keywords:
+                if keyword.arg == 'mode':
+                    modes = _chmod_get_mode(keyword.value)
+                    break
+    except ValueError:
+        return False
+    else:
+        if modes is None:
+            # NB: this would be from invalid code such as `os.chmod("file.txt")`
+            raise RuntimeError('Unable to extract `mode` argument from function call!')
+        return bool(modes & (stat.S_IWGRP | stat.S_IXGRP | stat.S_IWOTH | stat.S_IXOTH))
+
+
+# ==============================================================================
+
 
 class SecureCodingStandardChecker(BaseChecker):  # pylint: disable=too-many-instance-attributes
     """Plugin class."""
@@ -429,6 +518,11 @@ class SecureCodingStandardChecker(BaseChecker):  # pylint: disable=too-many-inst
             'os-mknod-unsafe-permissions',
             'Avoid using `os.mknod` with unsafe file permissions (by default 0 <= mode <= 0o755)',
         ),
+        'W8019': (
+            'Avoid using `os.chmod` with unsafe permissions (W ^ X for group and others)',
+            'os-chmod-unsafe-permissions',
+            'Avoid using `os.chmod` with unsafe file permissions (W ^ X for group and others)',
+        ),
     }
 
     def __init__(self, linter):
@@ -479,6 +573,8 @@ class SecureCodingStandardChecker(BaseChecker):  # pylint: disable=too-many-inst
             self.add_message('avoid-marshal-load', node=node)
         elif _is_function_call(node, module='shelve', function='open'):
             self.add_message('avoid-shelve-open', node=node)
+        elif _is_function_call(node, module='os', function='chmod') and _chmod_has_wx_for_go(node):
+            self.add_message('os-chmod-unsafe-permissions', node=node)
         elif _is_unix():
             if (
                 _is_function_call(node, module='os', function=('mkdir', 'makedirs'))